Number formatter supports and uses round half away from zero.
This adds a setting to the number formatter, and indeed a different
default, to choose how numbers are rounded. I am breaking compatibility
because I strongly feel that this is a better default (least surprise
for users reading the visuzlization's numbers).

I have some reservations with the name of the method setting function,
and generally about how this is done. I think it would be better and
generally more useful to add an unbiased rounding implementation (round
half to even) to the pv space: `pv.round`.