Adding support for dashed lines.
All stroked paths now support dash patterns. This is mostly useful for rules and lines but for completeness I have added dashes support to all shapes.

Given that Protovis is a little inconsistent with it's property names I have picked `lineDash` to harmonize with other stroke attributes: `lineJoin` and `lineWidth`. This makes `strokeStyle` the odd one out. It would probably be better to rename that to `lineColor` and `fillStyle` to `fillColor`.